// packages/twenty-server/src/engine/metadata-modules/timeline-activity-type/enums/timeline-activity-type-exception-code.enum.ts
export enum TimelineActivityTypeExceptionCode {
  TIMELINE_ACTIVITY_TYPE_NOT_FOUND = 'TIMELINE_ACTIVITY_TYPE_NOT_FOUND',
  TIMELINE_ACTIVITY_TYPE_NAME_ALREADY_EXISTS = 'TIMELINE_ACTIVITY_TYPE_NAME_ALREADY_EXISTS',
  TIMELINE_ACTIVITY_TYPE_IS_STANDARD = 'TIMELINE_ACTIVITY_TYPE_IS_STANDARD',
  INVALID_TIMELINE_ACTIVITY_TYPE_INPUT = 'INVALID_TIMELINE_ACTIVITY_TYPE_INPUT',
}